requirements.txt
fastapi
uvicorn
python-multipart
requests
pdf2image
pillow
pydantic
openai 
python-dotenv